Refactor into facefind package with console scripts
