+++ facefind/__init__.py
@@ -1,0 +1,3 @@
+"""FaceFind package with shared utilities and CLI entry points."""
+
+__all__ = []
